Range padding
