+++ pmlc/dialect/tile/transforms/BUILD
@@ -26,7 +26,9 @@ plaidml_cc_library(
         "constant_types.cc",
         "contraction.cc",
         "make_program.cc",
-        "padding.cc",
+	"padding.cc",
+        "pad_ranges.cc",
+        "pad_constraints.cc",
         "pass_detail.h",
     ],
     hdrs = [
